Big changes
Remove tserv-service usage
Add wooting-ipc module, which is largely a stripped-down copy of tserv-service with the previous API baked-in to make things easier

The main reason for the change is simplicity. Don't need the API loading from tserv-service for this and it results in 200 less lines of code. There are some minor functional differences between wooting-ipc and tserv-service but for the most part it's the same. This was a combination of me being lazy and so I didn't have to change too much of the existing code that used it.